refactor: DAO를 사용해서 movie ticket을 저장하도록 변경

+++ app/src/main/java/woowacourse/movie/presentation/seats/SeatsContract.kt
@@ -14,7 +14,7 @@ interface SeatsContract {
 
         fun updateConfirmButton(enabled: Boolean)
 
-        fun moveToReservationResult(movieTicketId: Int)
+        fun moveToReservationResult(movieTicketId: Long)
 
         fun showMessage(message: String)
     }
